Added `updates.json`.
Flushed file with parts of the changelog, will need to continue to flush out so it contains entirety of the changelog and updates sections.

Instead of hard coding the content into the index using a database (considering difference between actual database and just continuing to use the JSON or XML format) it will allow for more flexibility and a lightweight homepage at the cost of speed when fetching content from the database.